Challenge Task 3
Since demo.war has an embedded Jetty, I have removed the tomcat install
and directly used java to start the server using systemd service.
This configuration is closer to a simple 'one process in one container'
mantra...

Demo is still running as root and need to test puppet dependencies,
for example does the server is restarted when the environment changes.