Challenge Task 5 (Cont.)
* Added a very simple healthcheck service

My view is that building this type of application healthcheck is
belongs to the application not the operation.  It is a question
of efficiency.  Service developers know more about the semantic
of the service they are developing and should take the responsibility
of assessing service health, which includes dependency health.

There are tools for this.  Dropwizard is a good example.
At Product Health, my previous company, we have developed a
simple health check handler (https://github.com/ProductHealth/gose4).
Equivalent tools exists for Python etc...